Moved examples into their own folder, started drafting inference
